+++ src/simple/FastaReducer.java
@@ -2,13 +2,13 @@ package simple;
 
 import java.io.IOException;
 
-import org.apache.hadoop.io.LongWritable;
+import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
-public class FastaReducer extends Reducer<LongWritable, Text, Text, Text> 
+public class FastaReducer extends Reducer<IntWritable, Text, Text, Text> 
 {
-	public void reduce(LongWritable key, Iterable<Text> values, Context context) throws IOException, InterruptedException 
+	public void reduce(IntWritable key, Iterable<Text> values, Context context) throws IOException, InterruptedException 
 	{
 		StringBuilder res = new StringBuilder("");
